adicionar testes do método da divisão

+++ src/hash/div.c
@@ -6,6 +6,6 @@
 
 ULL hash_div(ULL valor, int tam, int _)
 {
-  unsigned int hashed_value = valor % MOD;
+  ULL hashed_value = valor % MOD;
   return hashed_value % tam;
 }
